+++ include/highfive/bits/H5File_misc.hpp
@@ -78,7 +78,7 @@ inline File::File(const std::string& filename, unsigned openFlags,
 inline const std::string& File::getName() const noexcept {
     if (_filename.empty()) {
         _filename = details::get_name([this](char* buffer, hsize_t length) {
-            return H5Fget_name(getId(), buffer, length);
+            return H5Fget_name(getId(), buffer, static_cast<size_t>(length));
         });
     }
     return _filename;
